Extract RxCacheTest
