+++ airframe-rx/src/test/scala/wvlet/airframe/rx/RxTest.scala
@@ -728,95 +728,4 @@ object RxTest extends AirSpec {
     }
   }
 
-  test("cache") {
-    val v           = Rx.variable(1)
-    val rx: Rx[Int] = v.map(x => x * 10).cache
-    val c0 = RxRunner.runContinuously(rx) { e =>
-      e shouldBe OnNext(10)
-    }
-    c0.cancel
-
-    val events = Seq.newBuilder[RxEvent]
-    v := 2
-    val c1 = RxRunner.runContinuously(rx)(events += _)
-    c1.cancel
-    events.result() shouldBe Seq(
-      OnNext(10),
-      OnNext(20)
-    )
-    val events2 = Seq.newBuilder[RxEvent]
-    v := 3
-    val c2 = RxRunner.runContinuously(rx)(events2 += _)
-    v := 4
-    c2.cancel
-    events2.result() shouldBe Seq(
-      OnNext(20),
-      OnNext(30),
-      OnNext(40)
-    )
-  }
-
-  test("cache.expireAfterWrite") {
-    val ticker = Ticker.manualTicker
-    val v      = Rx.variable(1)
-    val rx     = v.map(x => x * 10).cache.expireAfterWrite(1, TimeUnit.HOURS).withTicker(ticker)
-    val c0 = RxRunner.runContinuously(rx) { e =>
-      e shouldBe OnNext(10)
-    }
-    c0.cancel
-    val events = Seq.newBuilder[RxEvent]
-    v := 2
-
-    val c1 = RxRunner.runContinuously(rx)(events += _)
-    c1.cancel
-    events.result() shouldBe Seq(
-      OnNext(10),
-      OnNext(20)
-    )
-
-    events.clear()
-    v := 3
-    // Force expiration of the cache
-    ticker.advance(1, TimeUnit.HOURS)
-    val c2 = RxRunner.runContinuously(rx)(events += _)
-    c2.cancel
-    events.result() shouldBe Seq(
-      OnNext(30)
-    )
-  }
-
-  test("cache for option") {
-    val v  = Rx.optionVariable(Some(1))
-    val rx = v.cache
-    val c1 = RxRunner.runContinuously(rx) { _ shouldBe OnNext(Some(1)) }
-    c1.cancel
-
-    v := Some(2)
-    val events = Seq.newBuilder[RxEvent]
-    val c2     = RxRunner.runContinuously(rx)(events += _)
-    events.result() shouldBe Seq(
-      OnNext(Some(1)),
-      OnNext(Some(2))
-    )
-    c2.cancel
-  }
-
-  test("cache expiration for option") {
-    val ticker = Ticker.manualTicker
-    val v      = Rx.optionVariable(Some(1))
-    val rx     = v.cache.expireAfterWrite(1, TimeUnit.HOURS).withTicker(ticker)
-    val c1     = RxRunner.runContinuously(rx) { _ shouldBe OnNext(Some(1)) }
-    c1.cancel
-
-    v := Some(2)
-    // Force expiration of the cache
-    ticker.advance(1, TimeUnit.HOURS)
-    val events = Seq.newBuilder[RxEvent]
-    val c2     = RxRunner.runContinuously(rx)(events += _)
-    events.result() shouldBe Seq(
-      OnNext(Some(2))
-    )
-    c2.cancel
-  }
-
 }
